requirements.txt
Django==1.8.6
django-inspect
django-embed-video
django-bootstrap-form
django-redis
django-filter
